+++ internal/provider/data_source_iosxe_bgp_ipv4_unicast_vrf_neighbor.go
@@ -223,6 +223,10 @@ func (d *BGPIPv4UnicastVRFNeighborDataSource) Schema(ctx context.Context, req da
 				MarkdownDescription: "",
 				Computed:            true,
 			},
+			"ha_mode_graceful_restart": schema.BoolAttribute{
+				MarkdownDescription: "graceful-restart for this peer",
+				Computed:            true,
+			},
 		},
 	}
 }